Refactor folder structure & permissions
